Q1/Q2 finish

+++ programs/q1/src/cs455/hadoop/q1/Q1Job.java
@@ -2,7 +2,7 @@ package cs455.hadoop.q1;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -31,12 +31,12 @@ public class Q1Job {
             job.setReducerClass(Q1Reducer.class);
             // Outputs from the Mapper.
             job.setMapOutputKeyClass(Text.class);
-            job.setMapOutputValueClass(IntWritable.class);
+            job.setMapOutputValueClass(DoubleWritable.class);
             // Outputs from Reducer. It is sufficient to set only the following two properties
             // if the Mapper and Reducer has same key and value types. It is set separately for
             // elaboration.
             job.setOutputKeyClass(Text.class);
-            job.setOutputValueClass(IntWritable.class);
+            job.setOutputValueClass(DoubleWritable.class);
             // path to input in HDFS
             FileInputFormat.addInputPath(job, new Path("/data/main"));
             // path to output in HDFS
